Move the exit quiz confirmation modal to shared component and clear selected answer of the multi-intelligent test
